+++ src/Absences.php
@@ -8,7 +8,7 @@ use Swe\SpaceSDK\Exception\MissingArgumentException;
 
 /**
  * Class Absences
- * Generated at 2023-01-27 02:00
+ * Generated at 2023-02-07 02:00
  *
  * @package Swe\SpaceSDK
  * @author Luca Braun <l.braun@s-w-e.com>