+++ src/Administration/UserAgreement/Enabled.php
@@ -9,7 +9,7 @@ use Swe\SpaceSDK\Type;
 
 /**
  * Class Enabled
- * Generated at 2023-01-27 02:00
+ * Generated at 2023-02-07 02:00
  *
  * @package Swe\SpaceSDK\Administration\UserAgreement
  * @author Luca Braun <l.braun@s-w-e.com>