+++ src/Calendars.php
@@ -13,7 +13,7 @@ use Swe\SpaceSDK\Calendars\NonWorkingDaysEvents;
 
 /**
  * Class Calendars
- * Generated at 2023-01-27 02:00
+ * Generated at 2023-02-07 02:00
  *
  * @package Swe\SpaceSDK
  * @author Luca Braun <l.braun@s-w-e.com>