+++ src/Chats/Channels.php
@@ -16,7 +16,7 @@ use Swe\SpaceSDK\Type;
 
 /**
  * Class Channels
- * Generated at 2023-01-27 02:00
+ * Generated at 2023-02-07 02:00
  *
  * @package Swe\SpaceSDK\Chats
  * @author Luca Braun <l.braun@s-w-e.com>