+++ src/CustomFields/Fields/EnumValues.php
@@ -9,7 +9,7 @@ use Swe\SpaceSDK\Type;
 
 /**
  * Class EnumValues
- * Generated at 2023-01-27 02:00
+ * Generated at 2023-02-07 02:00
  *
  * @package Swe\SpaceSDK\CustomFields\Fields
  * @author Luca Braun <l.braun@s-w-e.com>