+++ src/TeamDirectory/CalendarEvents/MeetingParticipations.php
@@ -9,7 +9,7 @@ use Swe\SpaceSDK\Type;
 
 /**
  * Class MeetingParticipations
- * Generated at 2023-01-27 02:00
+ * Generated at 2023-02-07 02:00
  *
  * @package Swe\SpaceSDK\TeamDirectory\CalendarEvents
  * @author Luca Braun <l.braun@s-w-e.com>